Added sensitivity to bat area

+++ ball.h
@@ -8,8 +8,8 @@ typedef struct Ball
 {
   // SDL_Rect srect;
   SDL_Rect drect;
-  int y_direction;
-  int x_direction;
+  double y_direction;
+  double x_direction;
 } Ball;
